+++ lib/js_regex.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # JsRegex converts ::Regexp instances to JavaScript.
 #
 # Usage: