+++ lib/js_regex/conversion.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class JsRegex
   #
   # This class acts as a facade, passing a Regexp to the Converters.