+++ spec/lib/js_regex/converter/escape_converter_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 #
 #
 #
@@ -141,17 +143,17 @@ describe JsRegex::Converter::EscapeConverter do
   end
 
   it 'converts control sequences to unicode escapes' do
-    expect(Regexp.new('\C-*'.force_encoding('ascii-8bit'))).to\
+    expect(Regexp.new('\C-*'.dup.force_encoding('ascii-8bit'))).to\
     become(/\u000A/).and keep_matching("ya\ny", with_results: %W[\n])
   end
 
   it 'converts meta sequences to unicode escapes' do
-    expect(Regexp.new('\M-X'.force_encoding('ascii-8bit'))).to\
+    expect(Regexp.new('\M-X'.dup.force_encoding('ascii-8bit'))).to\
     become(/\u00D8/)
   end
 
   it 'converts meta control sequences to unicode escapes' do
-    expect(Regexp.new('\M-\C-X'.force_encoding('ascii-8bit'))).to\
+    expect(Regexp.new('\M-\C-X'.dup.force_encoding('ascii-8bit'))).to\
     become(/\u0098/)
   end