+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [3.14.0] - 2025-08-03
+
 ### Added
 
 - enable frozen string literals (#42)